+++ x-pack/plugins/lens/public/indexpattern_datasource/operations/definitions/formula/editor/formula.scss
